feat: Implement optimizers module with natural gradient methods - Separate optimization into update direction and update rule components - Use optax for first-order optimizers (adam, sgd); implement natural gradient methods (SR, LM) with QGT-based directions - Compute QGT in engine.geometry with 'matvec' and 'dense' backends, sharing Jacobian via jax.linearize and jax.linear_transpose - Initial SR testing with RBM model reveals need for numerical stability and local minima improvements - LM, Amoji line search, and trust region methods planned for future implementation

+++ lever/models/__init__.py
@@ -192,7 +192,7 @@ def RBMModPhase(
     """
     from .rbm import RBMModPhase
 
-    ansatz = RBMModPhase(alpha=alpha, param_dtype=jnp.float64)
+    ansatz = RBMModPhase(alpha=alpha, param_dtype=jnp.float32)
     return make_model(module=ansatz, seed=seed, n_orbitals=n_orbitals)
